Complete spacing for comments and add placeholders
